refactor(server): Move ZallyAssertions to zally-test #560
