+++ server/src/test/java/de/zalando/zally/rule/zally/TagAllOperationsRuleTest.kt
@@ -1,7 +1,7 @@
 package de.zalando.zally.rule.zally
 
 import de.zalando.zally.rule.DefaultContextFactory
-import de.zalando.zally.rule.ZallyAssertions
+import de.zalando.zally.test.ZallyAssertions
 import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Violation
 import org.intellij.lang.annotations.Language
