+++ server/zally-test/src/main/kotlin/de/zalando/zally/test/ZallyAssertions.kt
@@ -1,4 +1,4 @@
-package de.zalando.zally.rule
+package de.zalando.zally.test
 
 import de.zalando.zally.rule.api.Violation
